chore: move getDecryptedVault and setEncryptedVault to utils
