+++ packages/sysweb3-utils/src/accounts.ts
@@ -1,14 +1,24 @@
-export type IEthereumAddress = {
-  address: IEthereumBalance[];
+import { sysweb3Di } from '@pollum-io/sysweb3-core';
+
+const storage = sysweb3Di.getStateStorageDb();
+
+export const setEncryptedVault = (decryptedVault: any) => {
+  const encryptedVault = CryptoJS.AES.encrypt(
+    JSON.stringify(decryptedVault),
+    storage.get('vault-keys').hash
+  );
+
+  storage.set('vault', encryptedVault.toString());
 };
 
-export type IEthereumBalance = {
-  balances: IEthereumCurrency[];
-};
+export const getDecryptedVault = () => {
+  const vault = storage.get('vault');
 
-export type IEthereumCurrency = {
-  currency: {
-    symbol: string;
-  };
-  value: number;
+  const { hash } = storage.get('vault-keys');
+
+  const decryptedVault = CryptoJS.AES.decrypt(vault, hash).toString(
+    CryptoJS.enc.Utf8
+  );
+
+  return JSON.parse(decryptedVault);
 };
