+++ packages/sysweb3-utils/src/index.ts
@@ -9,7 +9,6 @@ export * from './signer';
 export * from './accounts';
 export * from './networks';
 export * from './spsbt';
-export * from './signer-info';
 export * from './address';
 export * from './constants';
 export * from './rpc';
