--- packages/sysweb3-utils/src/signer-info.ts
@@ -1,6 +1,0 @@
-export type SignerInfo = {
-  walletMnemonic: string;
-  isTestnet: boolean;
-  network: string;
-  blockbookURL: string;
-};
